services/ai/requirements.txt
# Core web framework
fastapi==0.115.12
uvicorn==0.34.3
pydantic==2.11.5

# HTTP client
httpx==0.28.1

# LLM providers
anthropic==0.68.1
boto3==1.40.40  # For AWS Bedrock

# ML/AI libraries
torch==2.7.1
transformers==4.52.4
numpy==2.3.0

# LlamaIndex for document processing
llama-index==0.12.42
llama-index-embeddings-huggingface==0.5.4

# Document processing
pypdf==5.6.0

# Database
asyncpg==0.30.0

# Utilities
python-ulid==3.1.0

# Testing
pytest==8.4.0
requests==2.32.4  # For test HTTP requests
